+++ runtime/src/test/java/com/instana/android/BaseTest.kt
@@ -17,6 +17,7 @@ import org.robolectric.annotation.Config
 import java.io.File
 import java.lang.reflect.Method
 import java.nio.file.Files
+import java.time.Instant
 
 @RunWith(RobolectricTestRunner::class)
 @Config(minSdk = 31, maxSdk = 31)
@@ -63,14 +64,60 @@ abstract class BaseTest {
         return method.invoke(obj, modifiedParam)
     }
 
-    internal fun createListOfFiles(directoryPath: String, fileCount: Int): ArrayList<File> {
+    internal fun invokePrivateMethod3(
+        obj: Any,
+        methodName: String,
+        param1: Any?,
+        paramType1: Class<*>,
+        param2: Any?,
+        paramType2: Class<*>
+    ): Any? {
+        // Retrieve the method with the matching parameter types
+        val method: Method = obj.javaClass.getDeclaredMethod(methodName, paramType1, paramType2)
+
+        // Convert the first primitive parameter to its wrapper class
+        val modifiedParam1 = if (param1 != null && paramType1.isPrimitive) {
+            when (param1) {
+                is Int -> param1 as Int
+                is Long -> param1 as Long
+                is Double -> param1 as Double
+                // Add other primitive types as needed
+                else -> throw IllegalArgumentException("Unsupported primitive type: $paramType1")
+            }
+        } else {
+            param1
+        }
+
+        // Convert the second primitive parameter to its wrapper class
+        val modifiedParam2 = if (param2 != null && paramType2.isPrimitive) {
+            when (param2) {
+                is Int -> param2 as Int
+                is Long -> param2 as Long
+                is Double -> param2 as Double
+                // Add other primitive types as needed
+                else -> throw IllegalArgumentException("Unsupported primitive type: $paramType2")
+            }
+        } else {
+            param2
+        }
+
+        method.isAccessible = true
+        // Invoke the method with two parameters
+        return method.invoke(obj, modifiedParam1, modifiedParam2)
+    }
+
+
+    internal fun createListOfFiles(directoryPath: String, fileCount: Int,olderNumberOfFiles:Int=0,modifiedAtMin:Long=15L): ArrayList<File> {
         val directory = Files.createTempDirectory(directoryPath).toFile()
 
         val files = arrayListOf<File>()
-
+        val olderModifiedTime = Instant.now().minusSeconds(modifiedAtMin * 60).toEpochMilli()
         repeat(fileCount) { index ->
             val tempFile = createTempFile(directory, "testFile_$index", ".txt")
             Files.write(tempFile.toPath(), "Test content for file $index  :  ".toByteArray())
+            if(olderNumberOfFiles!=0 && index<olderNumberOfFiles){
+                tempFile.setLastModified(olderModifiedTime)
+            }
             files.add(tempFile)
         }